667. Beautiful Arrangement II
