+++ src/main/kotlin/me/bossm0n5t3r/leetcode/beautifulArrangement2/README.md
